requirements.txt
fpdf2
selenium
webdriver-manager